Refactor(jiwon/market): Item 컴포넌트 분리 #26
